more lmp

+++ src/engine/search/quiescence.rs
@@ -15,10 +15,10 @@ pub fn quiescence(
     state: &mut SearchState,
     control: &impl Control,
 ) -> Result<Eval, ()> {
-    state.max_depth_reached = state.max_depth_reached.max(plies);
+    state.max_depth_reached = state.max_depth_reached.max(plies as i8);
     state.nodes_visited += 1;
 
-    if plies == MAX_SEARCH_DEPTH {
+    if plies == MAX_SEARCH_DEPTH as u8 {
         return Ok(eval::eval(game));
     }
